# SwaggerVideotapesGalore.yaml
swagger: '2.0'
info:
  description: >-
    Simple user and video tape storing and management system. Has functionality to 'borrow' video tapes to users via relations between the two tapes. Has additional functionality of managing and storing user rating of video tapes and user recommendations of video tapes. API designed by Edda Steinunn Rúnarsdóttir for the assigment Project #1 in the course T-302-HÖNN for Reykjavik University.
  version: 1.0.0
  title: Videotapes Galore API
host: '127.0.0.1:8080'
basePath: /api/v1
...
schemes:
  - http
paths:
...
  '/friends/{id}':
    get:
      tags:
        - friends
      summary: Gets specific friend by id
      operationId: getFriend
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id of friend to get
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Friend'
        '400':
          description: invalid id supplied
        '404':
          description: friend not found
    delete:
      tags:
        - friends
      summary: Deletes friend by id
      description: Deletes friend by id
      operationId: deleteFriendById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id of friend to delete
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: invalid id supplied
        '404':
          description: friend not found
    put:
      tags:
        - friends
      summary: Update friend by ID
      description: ''
      operationId: updateGameById
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: id of friend to delete
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: friend object to update
          required: true
          schema:
            $ref: '#/definitions/FriendInput'
      responses:
        '200':
          description: successful operation
        '400':
          description: invalid id supplied
        '404':
          description: friend not found
...
  /ratings/users/{userId}:
      get:
        tags:
          - ratings
        summary: Gets all ratings by given user on tapes
        description: Gets all ratings by given user on tapes
        operationId: getUserRatings
        produces:
          - application/json
        parameters:
        - name: userId
          in: path
          description: Id of user viewing his or her ratings on all video tapes
          required: true
          type: integer
          format: int64
        responses:
          '200':
            description: successful operation
            schema:
              type: array
              items:
                $ref: '#/definitions/Rating'
          '400':
            description: invalid id supplied for user
          '404':
            description: user not found with id
  /ratings/users/{userId}/tapes/{tapeId}:
      get:
        tags:
          - ratings
        summary: Gets given user's rating for a specific tape
        description: Gets given user's rating for a specific tape
        operationId: getUserRatingByTapeId
        produces:
          - application/json
        parameters:
        - name: userId
          in: path
          description: Id of user viewing his or her rating on video tape
          required: true
          type: integer
          format: int64
        - name: tapeId
          in: path
          description: Id of video tape user is viewing his or her rating on
          required: true
          type: integer
          format: int64
        responses:
          '200':
            description: successful operation
            schema:
              $ref: '#/definitions/RatingInput'
          '400':
            description: invalid id supplied for video tape or user
          '404':
            description: video tape or user not found
      post:
        tags:
          - ratings
        summary: Add given user's rating for a specific tape
        description: Add given user's rating for a specific tape
        operationId: postUserRatingByTapeId
        produces:
          - application/json
        parameters:
        - name: userId
          in: path
          description: Id of user adding his or her rating on video tape
          required: true
          type: integer
          format: int64
        - name: tapeId
          in: path
          description: Id of video tape user is adding his or her rating on
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Rating of given video tape for user
          required: true
          schema:
            $ref: '#/definitions/RatingInput'
        responses:
          '200':
            description: successful operation
          '400':
            description: invalid id supplied for video tape or user
          '404':
            description: video tape or user not found
      put:
        tags:
          - ratings
        summary: Update given user's rating for a specific tape
        description: Update given user's rating for a specific tape
        operationId: putUserRatingByTapeId
        produces:
          - application/json
        parameters:
        - name: userId
          in: path
          description: Id of user updating his or her rating on video tape
          required: true
          type: integer
          format: int64
        - name: tapeId
          in: path
          description: Id of video tape user is updating his or her rating on
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Rating of given video tape for user
          required: true
          schema:
            $ref: '#/definitions/RatingInput'
        responses:
          '200':
            description: successful operation
          '400':
            description: invalid id supplied for video tape or user
          '404':
            description: user rating not found on video tape
      delete:
        tags:
          - ratings
        summary: Delete given user's rating for a specific tape
        description: Delete given user's rating for a specific tape
        operationId: deleteUserRatingByTapeId
        produces:
          - application/json
        parameters:
        - name: userId
          in: path
          description: Id of user deleting his or her rating on video tape
          required: true
          type: integer
          format: int64
        - name: tapeId
          in: path
          description: Id of video tape user is deleting his or her rating on
          required: true
          type: integer
          format: int64
        responses:
          '200':
            description: successful operation
          '400':
            description: invalid id supplied for video tape or user
          '404':
            description: user rating not found on video tape
...
definitions:
  Friend:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      address:
        type: string
      email:
        type: string
      telephone:
        type: string
  FriendInput:
    type: object
    properties:
      name:
        type: string
      address:
        type: string
      email:
        type: string
      telephone:
        type: string
...
  Rating:
    type: object
    properties:
      userId:
        type: integer
        format: int64
      tapeId:
        type: integer
        format: int64
      rating:
        type: integer
        minimum: 1
        maximum: 5
  RatingInput:
    type: object
    properties:
      rating:
        type: integer
        minimum: 1
        maximum: 5